made prometheus testing folder with interactive testing file. moved old random metric flux testing function over from ipsm_test
