+++ npm/iptm/iptm_test.go
@@ -3,7 +3,6 @@ package iptm
 import (
 	"os"
 	"testing"
-	"time"
 
 	"github.com/Azure/azure-container-networking/npm/metrics"
 
@@ -234,51 +233,5 @@ func TestMain(m *testing.M) {
 
 	iptMgr.Restore(util.IptablesConfigFile)
 
-	// messWithMetrics()
-
 	os.Exit(exitCode)
-}
-
-func messWithMetrics() {
-	go func() {
-		for {
-			metrics.Inc(metrics.NumPolicies)
-			time.Sleep(2 * time.Second)
-		}
-	}()
-
-	go func() {
-		for k := 0; k < 25; k++ {
-			for j := 0; j < 2*k; j++ {
-				metrics.Inc(metrics.NumIPSets)
-			}
-			time.Sleep(2 * time.Second)
-		}
-	}()
-
-	go func() {
-		for j := 0; j < 500; j += 2 {
-			for k := 0; k < 2; k++ {
-				metrics.Observe(metrics.AddPolicyExecTime, float64(2*k*j))
-				time.Sleep(time.Second * time.Duration((k+1)/2))
-			}
-			for k := 0; k < 3; k++ {
-				metrics.Observe(metrics.AddPolicyExecTime, float64(-k+j))
-				time.Sleep(time.Second * time.Duration(k/3))
-			}
-		}
-	}()
-
-	go func() {
-		for {
-			for k := 0; k < 2; k++ {
-				metrics.Observe(metrics.AddIPSetExecTime, float64(2*k))
-				time.Sleep(time.Second * time.Duration((k+1)/2))
-			}
-			for k := 0; k < 3; k++ {
-				metrics.Observe(metrics.AddIPSetExecTime, float64(-k))
-				time.Sleep(time.Second * time.Duration(k+1))
-			}
-		}
-	}()
 }
