+++ npm/metrics/testing/test-util.go
